requirements.txt
streamlit
Pillow
numpy
opencv-python-headless
ultralytics
pyyaml
IPython
